requirements.txt
pandas
numpy
matplotlib
reportlab
psycopg2-binary
dotenv
os